# localscript-openapi.yaml
openapi: 3.0.3
info:
  title: LocalScript API
  description: 
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      summary: Проверка готовности сервиса и локального Ollama
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - ollama_reachable
                  - model_ready
                  - model
                  - num_ctx
                  - num_predict
                  - batch
                  - parallel
                  - gpu_only
                properties:
                  status:
                    type: string
                    enum: [ok, degraded]
                  ollama_reachable:
                    type: boolean
                  model_ready:
                    type: boolean
                  model:
                    type: string
                  num_ctx:
                    type: integer
                  num_predict:
                    type: integer
                  batch:
                    type: integer
                  parallel:
                    type: integer
                  gpu_only:
                    type: boolean

  /generate:
    post:
      summary: Сгенерировать код по тексту задачи
      operationId: generate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                  description: Текст задачи на естественном языке
                context:
                  type: object
                  description: Опциональный JSON-контекст (например структура wf.vars / wf.initVariables)
                previous_code:
                  type: string
                  description: Опционально — предыдущая версия Lua для итеративной доработки
                feedback:
                  type: string
                  description: Опционально — текст обратной связи пользователя для доработки
            example:
              prompt: "Функция factorial(n) для n >= 0"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code:
                    type: string
                    description: Сгенерированный Lua-код
                  confidence_gate_triggered:
                    type: boolean
                    description: true, если достигнут лимит repair-loop и код возвращен с низкой уверенностью
                  repair_report:
                    type: object
                    description: Отчет confidence-gate (история ошибок, повторяющиеся паттерны, рекомендации)
              example:
                code: |
                  function factorial(n)
                    if n <= 1 then return 1 end
                    return n * factorial(n - 1)
                  end

  /refine:
    post:
      summary: Итеративно доработать Lua-код по обратной связи
      operationId: refine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, previous_code, feedback]
              properties:
                prompt:
                  type: string
                context:
                  type: object
                  description: Опциональный JSON-контекст (например wf.vars)
                previous_code:
                  type: string
                feedback:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code:
                    type: string
                    description: Сгенерированный Lua-код
                  confidence_gate_triggered:
                    type: boolean
                    description: true, если достигнут лимит repair-loop и код возвращен с низкой уверенностью
                  repair_report:
                    type: object
                    description: Отчет confidence-gate (история ошибок, повторяющиеся паттерны, рекомендации)

  /clarify:
    post:
      summary: Сформировать уточняющие вопросы перед генерацией
      operationId: clarify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                context:
                  type: object
                  description: Опциональный JSON-контекст задачи
                answers:
                  type: array
                  description: Ответы пользователя на вопросы из предыдущего /clarify
                  items:
                    type: object
                    required: [id, text]
                    properties:
                      id:
                        type: string
                      text:
                        type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status, questions, assumptions]
                properties:
                  status:
                    type: string
                    enum: [need_clarification, ready_to_generate]
                  questions:
                    type: array
                    items:
                      type: object
                      required: [id, text, why, blocking, options]
                      properties:
                        id:
                          type: string
                        text:
                          type: string
                        why:
                          type: string
                        blocking:
                          type: boolean
                        options:
                          type: array
                          items:
                            type: string
                  assumptions:
                    type: array
                    items:
                      type: string
                  merged_context:
                    type: object
                  confidence_gate_triggered:
                    type: boolean
                  repair_report:
                    type: object

  /generate-from-clarify:
    post:
      summary: Уточнить требования и сразу сгенерировать код при достаточном контексте
      operationId: generateFromClarify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                context:
                  type: object
                answers:
                  type: array
                  items:
                    type: object
                    required: [id, text]
                    properties:
                      id:
                        type: string
                      text:
                        type: string
                previous_code:
                  type: string
                feedback:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status, questions, assumptions]
                properties:
                  status:
                    type: string
                    enum: [need_clarification, generated]
                  code:
                    type: string
                    nullable: true
                  questions:
                    type: array
                    items:
                      type: object
                      required: [id, text, why, blocking, options]
                      properties:
                        id:
                          type: string
                        text:
                          type: string
                        why:
                          type: string
                        blocking:
                          type: boolean
                        options:
                          type: array
                          items:
                            type: string
                  assumptions:
                    type: array
                    items:
                      type: string
                  merged_context:
                    type: object